bot/requirements.txt
aiohttp>=3.9
asyncssh>=2.14
yt-dlp>=2024.1
python-dotenv>=1.0
aiofiles>=23.0
pytest>=8.0
pytest-asyncio>=0.23
aioresponses>=0.7
